Chore : List<TypeResponse>로 변경
