+++ src/main/java/com/jelly/zzirit/domain/item/dto/response/BrandResponse.java
@@ -1,4 +1,4 @@
-package com.jelly.zzirit.domain.item.dto;
+package com.jelly.zzirit.domain.item.dto.response;
 
 public record BrandResponse (
 	Long brandId,
