+++ src/main/java/com/jelly/zzirit/domain/item/service/QueryTypeService.java
@@ -1,9 +1,11 @@
 package com.jelly.zzirit.domain.item.service;
 
+import java.util.List;
+
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.jelly.zzirit.domain.item.dto.TypeResponses;
+import com.jelly.zzirit.domain.item.dto.response.TypeResponse;
 import com.jelly.zzirit.domain.item.repository.TypeRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -15,7 +17,9 @@ public class QueryTypeService {
 
 	private final TypeRepository typeRepository;
 
-	public TypeResponses getAll() {
-		return TypeResponses.from(typeRepository.findAll());
+	public List<TypeResponse> getAll() {
+		return typeRepository.findAll().stream()
+			.map(TypeResponse::from)
+			.toList();
 	}
 }
